support total in get_winners
